fix(settings): step-by-step instructions for subscribing to the calendar feed (Issue 411)
users were uncertain how to actually paste the feed url into apple/google/outlook —
add a collapsible "how to add this to your calendar" panel with platform-specific
steps and links to each vendor's official help page. extracts the calendar-feed
section into its own file along the way.
